refactor where server and client routes are declared
