+++ x-pack/plugins/files/public/files_client/files_client.ts
@@ -5,9 +5,24 @@
  * 2.0.
  */
 
+import * as qs from 'query-string';
 import type { HttpStart } from '@kbn/core/public';
 import type { FilesClient } from '../types';
-import { FILE_KIND_API_ROUTES_CLIENT } from '../../common/api_routes';
+import { FILES_API_BASE_PATH } from '../../common/api_routes';
+
+const apiRoutes = {
+  getCreateFileRoute: (fileKind: string) => `${FILES_API_BASE_PATH}/${fileKind}`,
+  getUploadRoute: (fileKind: string, id: string) => `${FILES_API_BASE_PATH}/${fileKind}/${id}/blob`,
+  getDownloadRoute: (fileKind: string, id: string, fileName?: string) =>
+    `${FILES_API_BASE_PATH}/${fileKind}/${id}/blob/${fileName ? fileName : ''}`,
+  getUpdateRoute: (fileKind: string, id: string) => `${FILES_API_BASE_PATH}/${fileKind}/${id}`,
+  getDeleteRoute: (fileKind: string, id: string) => `${FILES_API_BASE_PATH}/${fileKind}/${id}`,
+  getListRoute: (fileKind: string, page?: number, perPage?: number) => {
+    const qParams = qs.stringify({ page, perPage });
+    return `${FILES_API_BASE_PATH}/${fileKind}/list${qParams ? `?${qParams}` : ''}`;
+  },
+  getByIdRoute: (fileKind: string, id: string) => `${FILES_API_BASE_PATH}/${fileKind}/${id}`,
+};
 
 interface Args {
   fileKind: string;
@@ -17,7 +32,7 @@ interface Args {
 export const createFilesClient = ({ http, fileKind }: Args): FilesClient => {
   return {
     create(args) {
-      return http.post(FILE_KIND_API_ROUTES_CLIENT.getCreateFileRoute(fileKind), {
+      return http.post(apiRoutes.getCreateFileRoute(fileKind), {
         headers: {
           'content-type': 'application/json',
         },
@@ -25,21 +40,19 @@ export const createFilesClient = ({ http, fileKind }: Args): FilesClient => {
       });
     },
     delete(args) {
-      return http.delete(FILE_KIND_API_ROUTES_CLIENT.getDeleteRoute(fileKind, args.id));
+      return http.delete(apiRoutes.getDeleteRoute(fileKind, args.id));
     },
     download(args) {
-      return http.get(
-        FILE_KIND_API_ROUTES_CLIENT.getDownloadRoute(fileKind, args.id, args.fileName)
-      );
+      return http.get(apiRoutes.getDownloadRoute(fileKind, args.id, args.fileName));
     },
     getById(args) {
-      return http.get(FILE_KIND_API_ROUTES_CLIENT.getByIdRoute(fileKind, args.id));
+      return http.get(apiRoutes.getByIdRoute(fileKind, args.id));
     },
     list(args) {
-      return http.get(FILE_KIND_API_ROUTES_CLIENT.getListRoute(fileKind, args.page, args.perPage));
+      return http.get(apiRoutes.getListRoute(fileKind, args.page, args.perPage));
     },
     update({ id, ...body }) {
-      return http.patch(FILE_KIND_API_ROUTES_CLIENT.getUpdateRoute(fileKind, id), {
+      return http.patch(apiRoutes.getUpdateRoute(fileKind, id), {
         headers: {
           'content-type': 'application/json',
         },
@@ -47,7 +60,7 @@ export const createFilesClient = ({ http, fileKind }: Args): FilesClient => {
       });
     },
     upload(args) {
-      return http.put(FILE_KIND_API_ROUTES_CLIENT.getUploadRoute(fileKind, args.id), {
+      return http.put(apiRoutes.getUploadRoute(fileKind, args.id), {
         body: args.body,
       });
     },
