+++ x-pack/plugins/files/server/routes/api_routes.ts
@@ -1,0 +1,25 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { FILES_API_BASE_PATH } from '../../common/api_routes';
+
+export * from '../../common/api_routes';
+
+export const FILES_API_ROUTES = {
+  find: `${FILES_API_BASE_PATH}/find`,
+  metrics: `${FILES_API_BASE_PATH}/metrics`,
+  fileKind: {
+    getCreateFileRoute: (fileKind: string) => `${FILES_API_BASE_PATH}/${fileKind}`,
+    getUploadRoute: (fileKind: string) => `${FILES_API_BASE_PATH}/${fileKind}/{id}/blob`,
+    getDownloadRoute: (fileKind: string) =>
+      `${FILES_API_BASE_PATH}/${fileKind}/{id}/blob/{fileName?}`,
+    getUpdateRoute: (fileKind: string) => `${FILES_API_BASE_PATH}/${fileKind}/{id}`,
+    getDeleteRoute: (fileKind: string) => `${FILES_API_BASE_PATH}/${fileKind}/{id}`,
+    getListRoute: (fileKind: string) => `${FILES_API_BASE_PATH}/${fileKind}/list`,
+    getByIdRoute: (fileKind: string) => `${FILES_API_BASE_PATH}/${fileKind}/{id}`,
+  },
+};
